pyproject.toml
[project]
name = "zotwatch"
version = "0.2.0"
description = "Personalized academic paper recommendation system based on Zotero library"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31",
    "pydantic>=2.6",
    "pyyaml>=6.0.1",
    "rapidfuzz>=3.5",
    "jinja2>=3.1",
    "feedparser>=6.0",
    "voyageai>=0.3",
    "faiss-cpu>=1.7",
    "numpy>=1.24",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[project.scripts]
zotwatch = "zotwatch.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zotwatch"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
docstring-code-format = true
